+++ esp-hal/src/clock/clocks_ll/esp32s2.rs
@@ -1,4 +1,8 @@
-use crate::clock::CpuClock;
+use crate::{
+    clock::CpuClock,
+    peripherals::SYSCON,
+    system::{RadioClockController, RadioPeripherals},
+};
 
 const MHZ: u32 = 1000000;
 const UINT16_MAX: u32 = 0xffff;
@@ -41,4 +45,100 @@ pub(crate) fn set_cpu_clock(cpu_clock_speed: CpuClock) {
         let value = (((80 * MHZ) >> 12) & UINT16_MAX) | ((((80 * MHZ) >> 12) & UINT16_MAX) << 16);
         rtc_cntl.store5().modify(|_, w| w.scratch5().bits(value));
     }
+}
+
+// Mask for clock bits used by both WIFI and Bluetooth, bit 0, 3, 6, 7, 8, 9
+const DPORT_WIFI_CLK_WIFI_BT_COMMON_M: u32 = 0x000003c9;
+const DPORT_WIFI_CLK_WIFI_EN_M: u32 = 0x000007cf;
+
+impl RadioClockController for crate::peripherals::RADIO_CLK {
+    fn enable(&mut self, peripheral: RadioPeripherals) {
+        match peripheral {
+            RadioPeripherals::Phy => enable_phy(),
+            RadioPeripherals::Wifi => wifi_clock_enable(),
+        }
+    }
+
+    fn disable(&mut self, peripheral: RadioPeripherals) {
+        match peripheral {
+            RadioPeripherals::Phy => disable_phy(),
+            RadioPeripherals::Wifi => wifi_clock_disable(),
+        }
+    }
+
+    fn reset_mac(&mut self) {
+        reset_mac();
+    }
+
+    fn init_clocks(&mut self) {
+        init_clocks();
+    }
+
+    fn ble_rtc_clk_init(&mut self) {
+        // nothing for this target
+    }
+
+    fn reset_rpa(&mut self) {
+        // nothing for this target
+    }
+}
+
+fn enable_phy() {
+    // `periph_ll_wifi_bt_module_enable_clk_clear_rst`
+    SYSCON::regs()
+        .wifi_clk_en()
+        .modify(|r, w| unsafe { w.bits(r.bits() | DPORT_WIFI_CLK_WIFI_BT_COMMON_M) });
+}
+
+fn disable_phy() {
+    // `periph_ll_wifi_bt_module_disable_clk_set_rst`
+    SYSCON::regs()
+        .wifi_clk_en()
+        .modify(|r, w| unsafe { w.bits(r.bits() & !DPORT_WIFI_CLK_WIFI_BT_COMMON_M) });
+}
+
+fn wifi_clock_enable() {
+    // `periph_ll_wifi_module_enable_clk_clear_rst`
+    SYSCON::regs()
+        .wifi_clk_en()
+        .modify(|r, w| unsafe { w.bits(r.bits() | DPORT_WIFI_CLK_WIFI_EN_M) });
+}
+
+fn wifi_clock_disable() {
+    // `periph_ll_wifi_module_disable_clk_set_rst`
+    SYSCON::regs()
+        .wifi_clk_en()
+        .modify(|r, w| unsafe { w.bits(r.bits() & !DPORT_WIFI_CLK_WIFI_EN_M) });
+}
+
+fn reset_mac() {
+    const SYSTEM_MAC_RST: u32 = 1 << 2;
+    SYSCON::regs()
+        .wifi_rst_en()
+        .modify(|r, w| unsafe { w.wifi_rst().bits(r.wifi_rst().bits() | SYSTEM_MAC_RST) });
+    SYSCON::regs()
+        .wifi_rst_en()
+        .modify(|r, w| unsafe { w.wifi_rst().bits(r.wifi_rst().bits() & !SYSTEM_MAC_RST) });
+}
+
+fn init_clocks() {
+    const DPORT_WIFI_CLK_WIFI_EN: u32 = 0x003807cf;
+    const DPORT_WIFI_CLK_BT_EN_M: u32 = 0x61 << 11;
+    const DPORT_WIFI_CLK_SDIOSLAVE_EN: u32 = 1 << 4;
+    const DPORT_WIFI_CLK_UNUSED_BIT5: u32 = 1 << 5;
+    const DPORT_WIFI_CLK_UNUSED_BIT12: u32 = 1 << 12;
+    const DPORT_WIFI_CLK_SDIO_HOST_EN: u32 = 1 << 13;
+    const DPORT_WIFI_CLK_EMAC_EN: u32 = 1 << 14;
+
+    const WIFI_BT_SDIO_CLK: u32 = DPORT_WIFI_CLK_WIFI_EN
+        | DPORT_WIFI_CLK_BT_EN_M
+        | DPORT_WIFI_CLK_UNUSED_BIT5
+        | DPORT_WIFI_CLK_UNUSED_BIT12
+        | DPORT_WIFI_CLK_SDIOSLAVE_EN
+        | DPORT_WIFI_CLK_SDIO_HOST_EN
+        | DPORT_WIFI_CLK_EMAC_EN;
+
+    SYSCON::regs()
+        .wifi_clk_en()
+        .modify(|r, w| unsafe { w.bits(r.bits() & !WIFI_BT_SDIO_CLK | DPORT_WIFI_CLK_WIFI_EN) });
 }
