+++ esp-hal/src/soc/esp32c2/mod.rs
@@ -7,7 +7,6 @@
 
 crate::unstable_module! {
     pub mod efuse;
-    pub mod radio_clocks;
     pub mod trng;
 }
 pub mod gpio;
