+++ esp-hal/src/soc/esp32c3/mod.rs
@@ -11,7 +11,6 @@
 
 crate::unstable_module! {
     pub mod efuse;
-    pub mod radio_clocks;
     pub mod trng;
 }
 pub mod gpio;
