+++ esp-hal/src/soc/esp32c6/mod.rs
@@ -13,7 +13,6 @@
 crate::unstable_module! {
     pub mod efuse;
     pub mod lp_core;
-    pub mod radio_clocks;
     pub mod trng;
 }
 pub mod gpio;
