+++ esp-hal/src/soc/esp32h2/mod.rs
@@ -12,7 +12,6 @@
 
 crate::unstable_module! {
     pub mod efuse;
-    pub mod radio_clocks;
     pub mod trng;
 }
 pub mod gpio;
