+++ esp-hal/src/soc/esp32s3/mod.rs
@@ -17,7 +17,6 @@ crate::unstable_module! {
     pub mod efuse;
     #[cfg(feature = "psram")]
     pub mod psram;
-    pub mod radio_clocks;
     pub mod trng;
     pub mod ulp_core;
 }
